+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ===
 
-* [feature] added `no-bare-expression` rule which disallows bare expression statements
+* [feature] added `no-unused-expression` rule which disallows unused expression statements
 * [bug] semicolon at end of file no longer triggers false positives for `whitespace` rule
 * [bug] hoisted functions no longer cause false positives for the `no-unreachable` rule
 * [bug] the rule loader no longer transforms/ignores the leading and trailing underscores and dashes of rule names in the config file